+++ .gitattributes
@@ -1,1 +1,2 @@
 * text=auto
+*.verified.cs text eol=lf working-tree-encoding=UTF-8
